% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_code.R
\name{capture_BiocCheck_output}
\alias{capture_BiocCheck_output}
\title{This function captures BiocCheck output}
\usage{
capture_BiocCheck_output(package)
}
\description{
This function captures BiocCheck output
}
