% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_issues_assignee}
\alias{get_issues_assignee}
\title{Download packages assigned to core-developer}
\usage{
get_issues_assignee(assignee)
}
\arguments{
\item{assignee}{character indicating the github username of
core-team member}
}
\description{
This function only works when a valid github username and password
are supplied to the function BiocPackageReview:::.github_userpwd()
}
\examples{
get_issue_assignee(assignee="nturaga")

}
